+++ NAMESPACE
@@ -9,11 +9,9 @@ export(KernelRectangular)
 export(Madscan)
 export(Meanscan)
 export(RRI)
-export(RRIcpp)
 export(Trik2)
 export(circularDispersionGV)
 export(circularDispersionNV)
 import(terra)
-importFrom(Rcpp,sourceCpp)
 importFrom(terra,lapp)
 useDynLib(SurfRough, .registration = TRUE)
