+++ R/RcppExports.R
@@ -1,7 +1,5 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
-
 RRI_cpp <- function(x, ni, nw) {
     .Call(`_SurfRough_RRI_cpp`, x, ni, nw)
 }
-
