+++ R/package.R
@@ -1,3 +1,2 @@
 #' @useDynLib SurfRough, .registration = TRUE
-#' @importFrom Rcpp sourceCpp
 NULL
